requirement.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-mistralai
mistralai

chromadb

sentence-transformers

python-dotenv

tiktoken

numpy
pandas 

fastapi
uvicorn

streamlit

lxml
beautifulsoup4
unstructured
pypdf

tqdm
requests





